refactor : joinCode 패키지 dto로 변경
객체로 매핑시켜 저장하는 역할만 하기 때문에 위치 변경
