+++ src/main/java/com/gagoo/thiscoding/domain/maria/user/controller/port/CertificationService.java
@@ -1,6 +1,6 @@
 package com.gagoo.thiscoding.domain.maria.user.controller.port;
 
-import com.gagoo.thiscoding.domain.maria.user.domain.JoinCode;
+import com.gagoo.thiscoding.domain.maria.user.domain.dto.JoinCode;
 
 public interface CertificationService {
     void sendJoinCode(String email);
